Improves I2P configuration usability
We now provide I2P functions that check whether the configuration that
the user provided is correct; specifically, we distinguish between a
connection refused and a protocol error (which means an existing but
invalid port was provided).

This paves the way for user friendly error messages in the web
interface.